Restructure v0_6 and fix tests
