+++ Cargo.toml
@@ -11,14 +11,14 @@ license = "MIT"
 exclude = [".*"]
 
 [features]
-shp = ["shapefile"]
-fgb = ["flatgeobuf"]
+default = ["flatgeobuf", "geojson", "shapefile", "geozero"]
 
 [dependencies]
 geo-types = { version = "0.7.11" }
 serde = { version = "1.0.188" }
 geoserde_derive = { version = "0.1.0", path = "derive" }
 geojson = { version = "0.24.2", optional = true }
+geozero = { version = "0.14.0", optional = true } # for v0_5
 shapefile = { version = "0.6.0", features = ["geo-types"], optional = true }
 dbase = { version = "0.5.0", features = ["serde"] }
 geo-traits = { version = "0.2.0" }
@@ -26,7 +26,6 @@ flatgeobuf = { version = "=4.6.0", default-features = false, optional = true }
 
 [dev-dependencies]
 anyhow = { version = "1.0.75", features = ["backtrace"] }
-dbase = { version = "0.5.0", features = ["serde"] }
 flatgeobuf = { version = "=4.6.0", default-features = false }
 geo-traits = { version = "0.2.0" }
 geo-types = { version = "0.7.11", features = ["serde"] }
