+++ src/v0_6/fgb/prop2_.rs
