--- src/v2/de/fmt/fgb/mod.rs
@@ -1,4 +1,0 @@
-mod feat;
-mod prop;
-
-pub use feat::FeatureParser;
