--- src/v2/de/fmt/geozero/mod.rs
@@ -1,5 +1,0 @@
-mod map;
-mod prop;
-
-use map::MapAdapter;
-pub use prop::PropertiesAdapter;
