--- src/v2/mod.rs
@@ -1,2 +1,0 @@
-pub mod de;
-pub mod ser;
