--- src/v2/ser.rs
@@ -1,5 +1,0 @@
-pub trait FormatProperty {
-    // geoserdeはRust構造体寄りだから、integerとかではなくi32
-    // 同じキーで複数回呼ばれるかもしれない（入れ子で同名フィールド）
-    fn format_i32(&self, key: &str, value: i32);
-}
