+++ tests/shapefile.rs
@@ -1,10 +1,12 @@
+#![cfg(feature="shapefile")]
+
 use std::{io::Cursor, vec};
 
 use geoserde::GeoDeserialize;
 use serde::{Deserialize, Serialize};
 
 #[test]
-fn shape_test() -> anyhow::Result<()> {
+fn de() -> anyhow::Result<()> {
     // Create in-memory shapefile
     let mut shp_buf = vec![];
     let mut dbf_buf = vec![];
@@ -20,12 +22,13 @@ fn shape_test() -> anyhow::Result<()> {
     }
 
     // Initialize reader
-    let mut shp_r = shapefile::ShapeReader::new(Cursor::new(shp_buf))?;
-    let mut dbf_r = shapefile::dbase::Reader::new(Cursor::new(dbf_buf))?;
+    let shp_r = shapefile::ShapeReader::new(Cursor::new(shp_buf))?;
+    let dbf_r = shapefile::dbase::Reader::new(Cursor::new(dbf_buf))?;
     let mut reader = shapefile::Reader::new(shp_r, dbf_r);
+
     // ジオメトリは1コピーとなる
     for res in reader.iter_shapes_and_records_as::<shapefile::Polyline, MyProperty>() {
-        let (geom, prop) = res.unwrap();
+        let (geom, _prop) = res.unwrap();
         let _geom = geo_types::MultiLineString::from(geom)
             .0
             .into_iter()
@@ -40,7 +43,7 @@ fn polyline(i: i32) -> shapefile::Polyline {
     geo_types::LineString::from(vec![[org, org + 0.1], [org + 0.2, org + 0.3]]).into()
 }
 
-#[derive(Serialize, Deserialize, GeoDeserialize)]
+#[derive(Serialize, Deserialize)]
 struct MyProperty {
     number: i32,
     text: String,
@@ -58,4 +61,12 @@ impl MyProperty {
             text: "two".into(),
         }
     }
+}
+
+#[derive(GeoDeserialize)]
+struct MyFeature {
+    _number: i32,
+    _text: String,
+    #[geometry]
+    _shape: geo_types::LineString,
 }
