+++ tests/v0_6/flatgeobuf.rs
@@ -1,17 +1,24 @@
-use flatgeobuf::{FallibleStreamingIterator, FeatureProperties};
-use geoserde::v2::de::{fmt::fgb::FeatureParser, DeserializeFeature, ParseFeature};
-use geozero::ToGeo;
+#![cfg(feature = "flatgeobuf")]
+
+use flatgeobuf::FallibleStreamingIterator;
+use geoserde::v0_6::{fgb::FeatureParser, DeserializeFeature, ParseFeature};
 
 #[test]
 fn test_parse_fgb() {
     let mut fgb_file = std::io::Cursor::new(include_bytes!("sample/a.fgb"));
     let fgb = flatgeobuf::FgbReader::open(&mut fgb_file).unwrap();
+    let head = fgb.header();
+    // .columns()
+    // .unwrap()
+    // .into_iter()
+    // .map(|x| (x.type_(), x.name()))
+    // .collect::<Vec<_>>();
     let mut feat_iter = fgb.select_all().unwrap();
-    let head = feat_iter.header().columns().unwrap().into_iter().map(|x| (x.type_(), x.name())).collect::<Vec<_>>();
     // 所有権エラー
     while let Some(feat) = feat_iter.next().unwrap() {
-        let point = MyFeature::deserialize_feature(FeatureParser::new(head, feat));
-        dbg!(point);
+        let point = MyFeature::deserialize_feature(FeatureParser::new(feat));
+        // let prop_0 = feat.property_n(0)?;
+
         // let _geom = feat.to_geo().unwrap();
         // let _prop = feat.property::<i32>("number").unwrap();
     }
@@ -25,7 +32,7 @@ fn test_parse_fgb_only_geom() {
         .select_all()
         .unwrap();
     while let Some(feat) = reader.next().unwrap() {
-        let point = geo_types::Point::deserialize_feature(feat);
+        let point = geo_types::Point::deserialize_feature(FeatureParser::new(feat));
         dbg!(point);
     }
 }
