+++ tests/v0_6/mod.rs
@@ -1,4 +1,4 @@
 mod csv;
 mod duplicate_field;
-// mod fgb; // FIXME
+mod flatgeobuf;
 mod fgb_nested;
